reto_backend/requirements.txt

asgiref==3.5.0
certifi==2021.10.8
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.4
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.1.0
PyJWT==2.4.0
pyparsing==3.0.7
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.5
